% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StalAge_1_0_modified.R
\name{age_model}
\alias{age_model}
\title{StalAge Age model 1.0}
\usage{
age_model(Daten, x_raw, Daten_orig)
}
\arguments{
\item{Daten}{Dating file; column names: age, error and depth}

\item{x_raw}{sample depths}
}
\value{
Returns the median age and 2.5 \% and 97.5 \% quantiles.
}
\description{
StalAge is an age modelling software developed by Denis Scholz & Dirk Hoffmann
Contact: Denis Scholz, Institute for Geosciences, University of Mainz, Becherweg 21, 55128 Mainz, Germany
Email: scholzd@uni-mainz.de
}
\references{
\insertRef{scholz2011stalage}{SISAL.AM}

\insertRef{R}{bibtex}
}
