% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model.R
\name{AM_SISAL}
\alias{AM_SISAL}
\title{Run age models for chosen entities.}
\usage{
AM_SISAL(prefix, run, file_path = file.path(getwd(), "SISAL_Age_Model"),
  b = T, bc = T, stalage = T, lI = T, lR = T, default = T)
}
\arguments{
\item{prefix}{prefix to SISAL files}

\item{run}{File specifying which entitis and age models for each entity to run, where to save the files and where the SISAL files are saved.}
}
\value{
Runs the age models.
}
\description{
Run age models for chosen entities.
}
