% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_link_entity.R
\docType{data}
\name{composite_link_entity}
\alias{composite_link_entity}
\title{composite_link_entity}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
composite_link_entity
}
\description{
csv file for SISAL database v1b; containing the composite entity and entities contained in the composite
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
