% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18O.R
\docType{data}
\name{d18O}
\alias{d18O}
\title{d18O}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
d18O
}
\description{
csv file for SISAL database v1b; containing the d18O isotope measurments with the corresponding depths
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
