% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating_lamina.R
\docType{data}
\name{dating_lamina}
\alias{dating_lamina}
\title{dating_lamina}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
dating_lamina
}
\description{
csv file for SISAL database v1b; containing the entity id's of the lamina dated entities
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
