% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_link_reference.R
\docType{data}
\name{entity_link_reference}
\alias{entity_link_reference}
\title{entity_link_reference}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
entity_link_reference
}
\description{
csv file for SISAL database v1b; containing the reference id for each entity_id
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
