% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_functions.R
\name{eval}
\alias{eval}
\title{Evaluate SISAL chronology.}
\usage{
eval(sisal_chrono, runFile, .dating = SISAL.AM::dating)
}
\arguments{
\item{sisal_chrono}{SISAL chrnology that is meant to be evaluted}

\item{runFile}{runFile containing the entity ids of the executed age models}

\item{.dating}{SISAL v1b dating csv file}
}
\value{
5 files: the evauated reversal check, the flexibility check, the 1st uncertainty check, the 2nd uncertainty check and a file containing the full evaluation of all 4 checks. Each passed check gives 1 point.
The more points a AM recieved the better the model.
}
\description{
Evaluates SISAL chronology following 4 checks:
1: necessary: No reversals (AM fails if this check fails)
2: necessary: Flexibility - AM has to folllow clear growth rate changes (AM fails if this checl fails)
3: sufficient: Uncertainty increases between dates and at hiatuses
4: sufficient: Increasing absolute uncertainty with age
}
