% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_SISAL}
\alias{filter_SISAL}
\title{Filter SISAL db for records of class I to IV.}
\usage{
filter_SISAL(m, dating. = SISAL.AM::dating, sample. = SISAL.AM::sample,
  entity. = SISAL.AM::entity, hiatus. = SISAL.AM::hiatus)
}
\arguments{
\item{m}{Number of class to filter for; m in[1,2,3,4]}
}
\value{
A vector of all records belonging to the chosen class.
}
\description{
class I: Records that run without any additional manipulation: at most tractable reversals, no hiatuses, U-series dates \\
class II: Records that at most consist of tractable reversals, only include U-series dates but have at least one hiatus \\
class III: Any record that is not U-series dated \\
class IV: Records that cannot be evaluated, since they have too few or poor inconsistent dates (i.e. non-tractable reversals) \\
}
