% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.R
\docType{data}
\name{gap}
\alias{gap}
\title{gap}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
gap
}
\description{
csv file for SISAL database v1b; containing the sample id and depth for each gap in the database
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
