% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bacon_median_quantile}
\alias{get_bacon_median_quantile}
\title{Determine median and quantiles for Bacon MC ensemble.}
\usage{
get_bacon_median_quantile(depth_eval, hiatus_tb, bacon_mcmc, q1 = 0.05,
  q2 = 0.95)
}
\arguments{
\item{depth_eval}{Sample depths for which interpolated age is wanted.}

\item{hiatus_tb}{Table containing the hiatus depths.}

\item{bacon_mcmc}{Bacon MC ensemble}

\item{q1}{quantile 1}

\item{q2}{quantile 2}
}
\value{
Table containing the sample depths, median ages and uncertainties.
}
\description{
Determine median and quantiles for Bacon MC ensemble.
}
