% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_lin_interp}
\alias{get_lin_interp}
\title{Linear interpolation for single MC simulation.}
\usage{
get_lin_interp(data, depth_eval, hiatus_tb)
}
\arguments{
\item{data}{Table containing single age ensemble and dating depths}

\item{depth_eval}{Sample depths.}

\item{hiatus_tb}{Table conatining hiatus depths and sample_id's}
}
\value{
Interpolated ages for sample depths.
}
\description{
Linear interpolation for single MC simulation.
}
