% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiatus.R
\docType{data}
\name{hiatus}
\alias{hiatus}
\title{hiatus}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
hiatus
}
\description{
csv file for SISAL database v1b; containing the sample id and depth for each record containing a hiatus
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
