% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lin_reg_no_hiatus}
\alias{lin_reg_no_hiatus}
\title{Interpolate ages using lin. reg. for single MC simulation; no hiatus.}
\usage{
lin_reg_no_hiatus(data, depth_eval)
}
\arguments{
\item{data}{Table containing dates and dating depths.}

\item{depth_eval}{Sample depths.}
}
\value{
Lin. regression fitted ages for sample depths.
}
\description{
Interpolate ages using lin. reg. for single MC simulation; no hiatus.
}
