% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear regression slopes and intecepts for single MC simulation}
\usage{
linear_regression(data, hiatus_tb)
}
\arguments{
\item{data}{Table containing single age ensemble and dating depths}

\item{hiatus_tb}{Table conatining hiatus depths and sample_id's}
}
\value{
Table containing slopes and interceptions for each section.
}
\description{
Linear regression slopes and intecepts for single MC simulation
}
