% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mc_ensemble}
\alias{mc_ensemble}
\title{Generate age ensembles depending on AM method.}
\usage{
mc_ensemble(linReg = F, linInterp = F, age, age_error, N,
  working_directory, file_name)
}
\arguments{
\item{linReg}{Logical. TRUE if MC ensemble is to be generated for lin. reg.}

\item{linInterp}{Logival. TRUE if MC ensemble is to be generated for lin. interp.}

\item{age}{Dates.}

\item{age_error}{Uncertainties to input dates.}

\item{N}{Number of interations.}
}
\value{
Lin. reg. ensemble of N interations for sample depths.
}
\description{
Generate age ensembles depending on AM method.
}
