% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mc_linInt}
\alias{mc_linInt}
\title{Generate lin. interp. ensemble.}
\usage{
mc_linInt(N, hiatus_tb, depth_dating, age_ensemble, depth_sample)
}
\arguments{
\item{N}{Number of interations.}

\item{hiatus_tb}{Table containing hiatus depths and hiatus sample_ids.}

\item{depth_dating}{Dating depths.}

\item{age_ensemble}{Mc ensemble of dating table.}

\item{depth_sample}{Sample depths.}
}
\value{
Lin. interp. ensemble of N interations for sample depths.
}
\description{
Generate lin. interp. ensemble.
}
