% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_functions.R
\name{merge_SISAL_chrono}
\alias{merge_SISAL_chrono}
\title{Build SISAL chronology.}
\usage{
merge_SISAL_chrono(runfile, .dating = SISAL.AM::dating,
  .sample = SISAL.AM::sample, .entity = SISAL.AM::entity)
}
\arguments{
\item{runFile}{Tabel containing entity ids and working_directory where age model chrnologies are saved.}

\item{entity}{SISAL v1b entity csv file}

\item{sample}{SISAL v1b sample csv file}

\item{dating}{SISAL v1b dating csv file}
}
\value{
SISAL chronolgy containg NA for the missing  age models
#' @example
m <- merge_sisal_chrono(runFile)
r <- as.data.frame(m[1])
s <- as.data.frame(m[2])
}
\description{
Builds the SISAL chronology for successfully executed age models. Supply runFile to method and chronology is build for the included entity ids.
}
