% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\docType{data}
\name{notes}
\alias{notes}
\title{notes}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
notes
}
\description{
csv file for SISAL database v1b; containing notes for each site
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
