% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_chronology.R
\docType{data}
\name{original_chronology}
\alias{original_chronology}
\title{original_chronology}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
original_chronology
}
\description{
csv file for SISAL database v1b; containing the original chronologies (if available) for each record
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
