% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_AM}
\alias{plot_AM}
\title{Plots successfully executed AM.}
\usage{
plot_AM(age, unknown, used, hiatuses, x_lim = 50, x_lim_bottom,
  y_lim = 600, steps_x = 10, steps_y = 100, cex_legend, color,
  file_name)
}
\arguments{
\item{age}{Table containing dates, uncertainies and dating depths of used ages.}

\item{unknown}{Table containing dates, uncertainies and dating depths of not used ages.}

\item{hiatuses}{Table containing hiatus depths.}

\item{x_lim}{Upper limit of x axis.}

\item{x_lim_bottom}{Lower limit of x axis.}

\item{y_lim}{Upper limit of y axis.}

\item{steps_x}{Ticks steps of x axis}

\item{steps_y}{Ticks steps of y axis.}

\item{color}{Color for each AM}

\item{file_name}{Entity_id and entity_name.}
}
\description{
Plots successfully executed AM.
}
