% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_IQR}
\alias{plot_IQR}
\title{Plots IQRs of successfully executed AM.}
\usage{
plot_IQR(iqr, depths, used, cex_legend, color_iqr, file_name)
}
\arguments{
\item{iqr}{Table containing IQRs of successfully executed AM.}

\item{depths}{Sample depths.}

\item{used}{Table containing dating depths.}

\item{color_iqr}{Color for each AM}

\item{file_name}{Entity_id and entity_name.}
}
\description{
Plots IQRs of successfully executed AM.
}
