% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_functions.R
\name{plot_sisal_eval}
\alias{plot_sisal_eval}
\title{Heatmap plot of the evaluated SISAL chronology.}
\usage{
plot_sisal_eval(sisal_eval)
}
\arguments{
\item{sisal_eval}{Evaluation for each entity and age model}
}
\value{
Plot.
}
\description{
Plots the evaluated chrnology with ggplot.
}
