% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotAMC}
\alias{plotAMC}
\title{Plot summary plot.}
\usage{
plotAMC(working_directory, file_name, b, bc, stalage, linreg, lininterp)
}
\arguments{
\item{working_directory}{File path from where data is loaded and plot is to be saved.}

\item{file_name}{Name of the folder for which the plot is to be generated.}

\item{b}{Logical. TRUE if Bacon is to be plotted.}

\item{bc}{Logical. TRUE if Bchron is to be plotted.}

\item{stalage}{Logical. TRUE if Stalage is to be plotted.}

\item{linreg}{Logical. TRUE if lin. reg. is to be plotted.}

\item{lininterp}{Logical. TRUE if lin. interp. is to be plotted.}
}
\value{
Plots all successfully age models and confidence intervals. Marks modified and artificial dates.
}
\description{
Plot summary plot.
}
