% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model.R
\name{run_SISAL_chrono}
\alias{run_SISAL_chrono}
\title{Run AM for single entity_id.}
\usage{
run_SISAL_chrono(runFile, entid, working_directory, bacon, bchron, stalage,
  linInterp, linReg, j.)
}
\arguments{
\item{entid}{entity_id}

\item{working_directory}{File path to where input files are saved.}

\item{bacon}{Logical. A parameter specifiying if Bacon AM is to be executed.}

\item{bchron}{Logical. A parameter specifiying if Bchron AM is to be executed.}

\item{stalage}{Logical. A parameter specifiying if StalAge AM is to be executed. #'}

\item{linInterp}{Logical. A parameter specifiying if lin. Interp. AM is to be executed.}

\item{linReg}{Logical. A parameter specifiying if lin. Reg. AM is to be executed.}

\item{data}{List of site_tb, dating_tb and ample_tb.}
}
\value{
Executes the specified age models, marks in a sepearte file the successfully and failed age models, writes an error file and plots the summary.
}
\description{
Run AM for single entity_id.
}
