% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model.R
\name{runBacon}
\alias{runBacon}
\title{Age model function for Bacon.}
\usage{
runBacon(working_directory, file_name, postbomb = 0, cc = 0)
}
\arguments{
\item{working_directory}{File path, where input files are saved.}

\item{file_name}{name of the entity_id.}

\item{N}{Number of iterations.}
}
\value{
Saves MC ensembel, bacon_chronology and AM plot.
}
\description{
Age model function for Bacon.
}
