% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model.R
\name{runLinInterp}
\alias{runLinInterp}
\title{Age model function for linear interpolation.}
\usage{
runLinInterp(working_directory, file_name, N = 2000)
}
\arguments{
\item{working_directory}{File path, where input files are saved.}

\item{file_name}{name of the entity_id.}

\item{N}{Number of iterations.}
}
\value{
Saves MC ensembel, lin_interp_chronology and AM plot.
}
\description{
Age model function for linear interpolation.
}
\references{
Telford, R. J. et al., Quaternary Science Reviews 23, 1-5 (2004)
}
