% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model.R
\name{runLinReg}
\alias{runLinReg}
\title{Age model function for linear regression.}
\usage{
runLinReg(working_directory, file_name, N = 2000)
}
\arguments{
\item{working_directory}{File path, where input files are saved.}

\item{file_name}{name of the entity_id.}

\item{N}{Number of iterations.}
}
\value{
Saves MC ensembel, lin_reg_chronology and AM plot.
}
\description{
Age model function for linear regression.
}
