% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_model.R
\name{runStalAge}
\alias{runStalAge}
\title{Age model function for StalAge}
\usage{
runStalAge(working_directory, file_name)
}
\arguments{
\item{working_directory}{File path, where input files are saved.}

\item{file_name}{name of the entity_id.}

\item{N}{Number of iterations.}
}
\value{
Saves MC ensembel, StalAge_chronology and AM plot.
}
\description{
Age model function for StalAge
}
