% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{sample}
\alias{sample}
\title{sample}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
sample
}
\description{
csv file for SISAL database v1b; containing the sample depths for each sample id
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
