% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StalAge_1_0_modified.R
\name{scan}
\alias{scan}
\title{Broadly scan dating file for reversals/outliers and increase error.}
\usage{
scan(depth, age, error)
}
\arguments{
\item{depth}{Dating depths}

\item{age}{Dates}

\item{error}{Age uncertainties for each date}
}
\value{
Modified dating file.
}
\description{
Broadly scan dating file for reversals/outliers and increase error.
}
\references{
Scholz, D. and Hoffmann, D. L., Quaternary Geochronology 6, 369-382 (2011)
}
