% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StalAge_1_0_modified.R
\name{scan_fine}
\alias{scan_fine}
\title{Fine scan dating input for reversals/outliers and increase errors at tagged depths.}
\usage{
scan_fine(Daten)
}
\arguments{
\item{dating_tb}{Table containing dates, errors and depths.}
}
\value{
Modified dating file.

 @references
\insertRef{scholz2011stalage}{SISAL.AM}

\insertRef{R}{bibtex}
}
\description{
Fine scan dating input for reversals/outliers and increase errors at tagged depths.
}
