% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scan_lin_interp}
\alias{scan_lin_interp}
\title{Braodly scan dating input for reversals/outliers and increase errors at tagged depths.}
\usage{
scan_lin_interp(dating_tb)
}
\arguments{
\item{dating_tb}{Table containing dates, errors and depths.}
}
\value{
Modified dating file.
}
\description{
#' This is a modified version of scan_fine() to work with the input data of the lin. interp.
}
\references{
\insertRef{scholz2011stalage}{SISAL.AM}

\insertRef{R}{bibtex}
}
