% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisal_chronology.R
\docType{data}
\name{sisal_chronology}
\alias{sisal_chronology}
\title{sisal_chronology}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
sisal_chronology
}
\description{
csv file for SISAL database v1b; containing AM for records with no available original chronology
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
