% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\docType{data}
\name{site}
\alias{site}
\title{site}
\format{csv files containing all available information}
\source{
\url{http://pastglobalchanges.org/science/wg/sisal/intro}
}
\usage{
site
}
\description{
csv file for SISAL database v1b; containing the site ids and site information for each record
}
\references{
Atsawawaranunt, K. et al and SISAL working group members, ESSD 10,1687-1713,(2018),

ComasBru, L. et al and SISAL working group members, CLIM PAST 15,1557-1579(2019)
}
\keyword{datasets}
