% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StalAge_1_0_modified.R
\name{slope}
\alias{slope}
\title{Determines slope of three adjacent dates to determine reverslas/outliers}
\usage{
slope(depth, age, error)
}
\arguments{
\item{depth}{Dating depths}

\item{age}{Dates}

\item{error}{Age uncertainties for each date}
}
\value{
Percentage of failed fits.
}
\description{
Determines slope of three adjacent dates to determine reverslas/outliers
}
\references{
\insertRef{scholz2011stalage}{SISAL.AM}

\insertRef{R}{bibtex}
}
